[HeterogeneousDWARF] Implement AMDGPU CFI, DebugInfo
Add -amdgpu-spill-cfi-saved-regs to enable complete debug info even when
codegen would otherwise not spill the EXEC mask. This is a workaround
until a solution to describe the EXEC mask without spills is integrated
based on the new structurizer.

Support emitting eh_frame for debug-info even when the target does not
support exceptions.

Extend MCAsmInfo to reflect whether a target opts in to the
heterogeneous DWARF extensions.

FIXME: Can this be split to another commit?
Teach DebugInfo about the new extension to permit testing.

Change-Id: Ic9ddc8a29d86a25a5ae5004e318cfa3e4d3ba164
Note: The encodings in BinaryFormat/Dwarf.def are subject to change.